Fixes unit tests and renames file with typo
