+++ src/constants.ts
@@ -1,3 +1,14 @@
-export const GAS_LIMIT = 500_000;
+// The number of past blocks to lookup when fetching events.
 export const BLOCK_COUNT_HISTORY_LIMIT = 300;
+
+// The default amount of time before a "retryable" promise is timed out and retried
 export const DEFAULT_RETRY_TIMEOUT_MS = 5_000;
+
+// The Priority Fee in Wei
+export const PRIORITY_FEE = 3120000000;
+
+// The Base Fee to Max Fee multiplier
+export const BASE_FEE_MULTIPLIER = 2;
+
+// The default gas limit for transactions
+export const GAS_LIMIT = 500_000;
