+++ src/evm/initialize-provider.ts
@@ -1,6 +1,5 @@
 import * as node from '@api3/airnode-node';
 import * as protocol from '@api3/airnode-protocol';
-import { LogsData } from '@api3/airnode-node';
 import { ethers } from 'ethers';
 import isNil from 'lodash/isNil';
 import { ChainConfig, EVMProviderState } from '../types';
@@ -11,7 +10,7 @@ const rrpBeaconServerAbi = new ethers.utils.Interface(protocol.RrpBeaconServerFa
 );
 
 //TODO: where to get abi from?
-const dapiServerAbi = [
+export const dapiServerAbi = [
   'function conditionPspBeaconUpdate(bytes32,bytes,bytes) view returns (bool)',
   'function fulfillPspBeaconUpdate(bytes32,address,address,address,uint256,bytes,bytes)',
 ];
@@ -24,7 +23,7 @@ const abis: { [contractName: string]: string | string[] } = {
 export const initializeProvider = async (
   chain: ChainConfig,
   providerUrl: string
-): Promise<LogsData<EVMProviderState | null>> => {
+): Promise<node.LogsData<EVMProviderState | null>> => {
   const provider = node.evm.buildEVMProvider(providerUrl, chain.id);
 
   const contracts = Object.entries(chain.contracts).reduce((acc, [contractName, contractAddress]) => {
